mm: Refactor paging module
