+++ Cargo.toml
@@ -3,6 +3,7 @@ resolver = "2"
 members = [
     "lib/armv9a",
     "lib/cose",
+    "lib/paging",
     "lib/rsi-el0",
     "lib/uart",
     "plat/fvp",
