+++ lib/paging/.cargo/config
@@ -1,0 +1,2 @@
+[build]
+target = "aarch64-unknown-none-softfloat"
