+++ plat/fvp/src/entry.rs
@@ -5,7 +5,7 @@ use armv9a::regs::*;
 use islet_rmm::config::{NUM_OF_CPU, RMM_STACK_SIZE};
 use islet_rmm::io::stdout;
 use islet_rmm::logger;
-use islet_rmm::rmm::translation;
+use islet_rmm::mm::translation;
 
 #[no_mangle]
 #[link_section = ".stack"]
