+++ rmm/Cargo.toml
@@ -11,6 +11,7 @@ lazy_static = { version = "1.4.0", features = ["spin_no_std"] }
 linked_list_allocator = "0.10.4"
 lock_api = "0.4.9"
 log = "0.4.17"
+paging = { path = "../lib/paging" }
 spin = "0.9.2"
 spinning_top = "0.2.4"
