--- rmm/armv9a/Cargo.toml
@@ -1,29 +1,0 @@
-[package]
-name = "armv9a"
-version = "0.0.1"
-authors = ["Islet Contributors"]
-edition = "2021"
-
-[lib]
-name = "armv9a"
-path = "src/lib.rs"
-
-[features]
-max_level_off   = ["log/max_level_off"  , "monitor/max_level_off"  ]
-max_level_error = ["log/max_level_error", "monitor/max_level_error"]
-max_level_warn  = ["log/max_level_warn" , "monitor/max_level_warn" ]
-max_level_info  = ["log/max_level_info" , "monitor/max_level_info" ]
-max_level_debug = ["log/max_level_debug", "monitor/max_level_debug"]
-max_level_trace = ["log/max_level_trace", "monitor/max_level_trace"]
-
-[dependencies]
-linked_list_allocator = "0.10.4"
-lock_api = "0.4.9"
-log = "0.4.17"
-monitor = { path = "../monitor" }
-spin = "0.9.2"
-spinning_top = "0.2.4"
-lazy_static = { version = "1.4.0", features = ["spin_no_std"] }
-
-[build-dependencies]
-cc = "1.0"
