+++ rmm/src/granule/mod.rs
@@ -1,15 +1,14 @@
 pub mod entry;
 pub mod translation;
 
-use crate::mm::address::PhysAddr;
-use crate::mm::error::Error;
-use crate::mm::page_table::{HasSubtable, Level};
-use crate::rmi::error::Error as RmiError;
-
-use entry::{Entry, Inner};
-use translation::{
+use self::entry::{Entry, Inner};
+use self::translation::{
     GranuleStatusTable, GRANULE_STATUS_TABLE, L0_TABLE_ENTRY_SIZE_RANGE, L1_TABLE_ENTRY_SIZE_RANGE,
 };
+use crate::rmi::error::Error as RmiError;
+use paging::address::PhysAddr;
+use paging::error::Error;
+use paging::page_table::{HasSubtable, Level};
 
 pub const GRANULE_SIZE: usize = 4096;
 
@@ -95,12 +94,12 @@ impl GranuleState {
 macro_rules! get_granule {
     ($addr:expr) => {{
         {
-            use crate::mm::address::PhysAddr;
-            use crate::mm::error::Error as MmError;
-            use crate::mm::page::Page;
-            use crate::mm::page_table::{self, PageTableMethods};
-            use crate::rmm::granule::translation::{GranuleSize, GRANULE_STATUS_TABLE};
-            use crate::rmm::granule::validate_addr;
+            use crate::granule::translation::{GranuleSize, GRANULE_STATUS_TABLE};
+            use crate::granule::validate_addr;
+            use paging::address::PhysAddr;
+            use paging::error::Error as MmError;
+            use paging::page::Page;
+            use paging::page_table::{self, PageTableMethods};
 
             if !validate_addr($addr) {
                 Err(MmError::MmInvalidAddr)
@@ -133,7 +132,7 @@ macro_rules! get_granule_if {
     ($addr:expr, $state:expr) => {{
         get_granule!($addr).and_then(|guard| {
             if guard.state() != $state {
-                use crate::mm::error::Error as MmError;
+                use paging::error::Error as MmError;
                 Err(MmError::MmStateError)
             } else {
                 Ok(guard)
@@ -148,7 +147,7 @@ macro_rules! get_granule_if {
 macro_rules! set_state_and_get_granule {
     ($addr:expr, $state:expr) => {{
         {
-            use crate::rmm::granule::set_granule_raw;
+            use crate::granule::set_granule_raw;
             set_granule_raw($addr, $state).and_then(|_| get_granule!($addr))
         }
     }};
@@ -245,10 +244,10 @@ pub fn to_rmi_result(res: Result<(), Error>) -> Result<(), RmiError> {
 
 #[cfg(test)]
 mod test {
-    use crate::mm::error::Error;
-    use crate::rmm::granule::translation::{GranuleStatusTable, GRANULE_STATUS_TABLE};
-    use crate::rmm::granule::{set_granule, GranuleState};
+    use crate::granule::translation::{GranuleStatusTable, GRANULE_STATUS_TABLE};
+    use crate::granule::{set_granule, GranuleState};
     use crate::set_state_and_get_granule;
+    use paging::error::Error;
 
     const TEST_ADDR: usize = 0x880c_0000;
     const TEST_WRONG_ADDR: usize = 0x7900_0000;
