+++ rmm/src/realm/mm/address.rs
@@ -1,10 +1,10 @@
 use core::fmt;
 use core::ops::{Add, AddAssign, BitAnd, BitOr, Sub, SubAssign};
 
-use crate::impl_addr;
-use crate::mm::page::Address;
+use paging::address::Address;
+use paging::impl_addr;
 
-pub use crate::mm::address::PhysAddr;
+pub use paging::address::PhysAddr;
 
 #[derive(Copy, Clone, PartialEq, Eq, PartialOrd, Ord)]
 pub struct GuestPhysAddr(usize);
