+++ rmm/src/realm/mm/stage2_tte.rs
@@ -1,4 +1,4 @@
-use crate::mm::address::PhysAddr;
+use paging::address::PhysAddr;
 
 use armv9a::{define_bitfield, define_bits, define_mask};
