+++ rmm/src/rmi/gpt.rs
@@ -1,12 +1,13 @@
 use crate::asm::{smc, SMC_SUCCESS};
 use crate::event::Mainloop;
+use crate::granule::{set_granule, GranuleState};
 use crate::listen;
-use crate::mm::error::Error as MmError;
+use crate::mm::page_table::{map, unmap};
 use crate::rmi;
 use crate::rmi::error::Error;
-use crate::rmm::granule::{set_granule, GranuleState};
-use crate::rmm::PageMap;
 use crate::{get_granule, get_granule_if, set_state_and_get_granule};
+
+use paging::error::Error as MmError;
 
 extern crate alloc;
 
@@ -15,16 +16,16 @@ const MARK_REALM: usize = 0xc400_01b0;
 const MARK_NONSECURE: usize = 0xc400_01b1;
 
 pub fn set_event_handler(mainloop: &mut Mainloop) {
-    listen!(mainloop, rmi::GRANULE_DELEGATE, |arg, _, rmm| {
-        mark_realm(rmm.mm, arg[0])
+    listen!(mainloop, rmi::GRANULE_DELEGATE, |arg, _, _| {
+        mark_realm(arg[0])
     });
 
-    listen!(mainloop, rmi::GRANULE_UNDELEGATE, |arg, _, rmm| {
-        mark_ns(rmm.mm, arg[0])
+    listen!(mainloop, rmi::GRANULE_UNDELEGATE, |arg, _, _| {
+        mark_ns(arg[0])
     });
 }
 
-fn mark_realm(mm: PageMap, addr: usize) -> Result<(), Error> {
+fn mark_realm(addr: usize) -> Result<(), Error> {
     let mut granule = match get_granule_if!(addr, GranuleState::Undelegated) {
         Err(MmError::MmNoEntry) => set_state_and_get_granule!(addr, GranuleState::Undelegated),
         other => other,
@@ -34,16 +35,16 @@ fn mark_realm(mm: PageMap, addr: usize) -> Result<(), Error> {
         return Err(Error::RmiErrorInput);
     }
 
-    mm.map(addr, true);
+    map(addr, true);
     set_granule(&mut granule, GranuleState::Delegated).map_err(|e| {
-        mm.unmap(addr);
+        unmap(addr);
         e
     })?;
-    mm.unmap(addr);
+    unmap(addr);
     Ok(())
 }
 
-fn mark_ns(mm: PageMap, addr: usize) -> Result<(), Error> {
+fn mark_ns(addr: usize) -> Result<(), Error> {
     let mut granule = get_granule_if!(addr, GranuleState::Delegated)?;
 
     if smc(MARK_NONSECURE, &[addr])[0] != SMC_SUCCESS {
@@ -53,11 +54,11 @@ fn mark_ns(mm: PageMap, addr: usize) -> Result<(), Error> {
         );
     }
 
-    mm.map(addr, false);
+    map(addr, false);
     set_granule(&mut granule, GranuleState::Undelegated).map_err(|e| {
-        mm.unmap(addr);
+        unmap(addr);
         e
     })?;
-    mm.unmap(addr);
+    unmap(addr);
     Ok(())
 }
