+++ rmm/src/rmi/realm/mod.rs
@@ -5,10 +5,11 @@ use self::params::Params;
 pub use self::rd::Rd;
 use super::error::Error;
 use crate::event::Mainloop;
+use crate::granule::{set_granule, GranuleState};
 use crate::host::pointer::Pointer as HostPointer;
 use crate::listen;
+use crate::mm::page_table;
 use crate::rmi;
-use crate::rmm::granule::{set_granule, GranuleState};
 use crate::{get_granule, get_granule_if};
 
 extern crate alloc;
@@ -29,9 +30,9 @@ pub fn set_event_handler(mainloop: &mut Mainloop) {
 
         let mut rd_granule = get_granule_if!(rd, GranuleState::Delegated)?;
         let rd_obj = rd_granule.content_mut::<Rd>();
-        rmm.mm.map(rd, true);
+        page_table::map(rd, true);
 
-        let params = copy_from_host_or_ret!(Params, params_ptr, rmm.mm);
+        let params = copy_from_host_or_ret!(Params, params_ptr);
         if params.rtt_base as usize == rd {
             return Err(Error::RmiErrorInput);
         }
@@ -50,7 +51,7 @@ pub fn set_event_handler(mainloop: &mut Mainloop) {
         };
 
         eplilog().map_err(|e| {
-            rmm.mm.unmap(rd);
+            page_table::unmap(rd);
             rmm.rmi.remove(id).expect("Realm should be created before.");
             e
         })
@@ -74,7 +75,7 @@ pub fn set_event_handler(mainloop: &mut Mainloop) {
 
         // change state when everything goes fine.
         set_granule(&mut rd_granule, GranuleState::Delegated)?;
-        rmm.mm.unmap(arg[0]);
+        page_table::unmap(arg[0]);
 
         Ok(())
     });
